Added F5 Functionality
Look for comments with !!! in game.py to see where I recommended adding the various things for Sprint 3

Note: Should consider moving the different draw screen function into a separate file/class as well as making the button inputs for the different active views in the file or class. Perhaps making a view.py class to handle how the screens are drawn as well how to handle button inputs on the different screens. This would help cleanup game.py.